dom/html/test/test_bug481647.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=481647
-->
<head>
  <title>Test for Bug 481647</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=481647">Mozilla Bug 481647</a>
<p id="display">
  <iframe src="javascript:'aaa'"></iframe>
  <iframe src="javascript:document.write('aaa'); document.close();"></iframe>
</p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 481647 **/
SimpleTest.waitForExplicitFinish()

function testFrame(num) {
  is(window.frames[num].document.baseURI, document.baseURI,
     "Unexpected base URI in frame " + num);
}

addLoadEvent(function() {
  for (var i = 0; i < 2; ++i) {
    testFrame(i);
  }

  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
